% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consumer.R
\name{Consumer}
\alias{Consumer}
\title{Consumer reference class}
\description{
Consumer reference class

Consumer reference class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Consumer-new}{\code{Consumer$new()}}
\item \href{#method-Consumer-subscribe}{\code{Consumer$subscribe()}}
\item \href{#method-Consumer-consume}{\code{Consumer$consume()}}
\item \href{#method-Consumer-commit}{\code{Consumer$commit()}}
\item \href{#method-Consumer-unsubscribe}{\code{Consumer$unsubscribe()}}
\item \href{#method-Consumer-close}{\code{Consumer$close()}}
\item \href{#method-Consumer-clone}{\code{Consumer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-new}{}}}
\subsection{Method \code{new()}}{
Create a new consumer instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$new(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{(`list()`)\cr
list of config values. These are passed without modification to
librdkafka}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
Subscribe to topic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$subscribe(topic)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topic}}{(`character`)\cr
kafka topic}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-consume"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-consume}{}}}
\subsection{Method \code{consume()}}{
Consume single message
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$consume(timeout = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{(`integer`)\cr
timeout in ms.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-commit"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-commit}{}}}
\subsection{Method \code{commit()}}{
Commit current offset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$commit(async = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{async}}{(`logical`)\cr
commit async}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-unsubscribe"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-unsubscribe}{}}}
\subsection{Method \code{unsubscribe()}}{
Unsubscribe from current topic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$unsubscribe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-close"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-close}{}}}
\subsection{Method \code{close()}}{
Close consumer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Consumer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Consumer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Consumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
