% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Message.R
\name{new_kafka_message}
\alias{new_kafka_message}
\title{Create KafkaMessage object}
\usage{
new_kafka_message(key = NULL, value)
}
\arguments{
\item{key}{(`character`) message key (optional)}

\item{value}{(`character`) message value}
}
\description{
Create KafkaMessage object
}
