% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Message.R
\name{print.KafkaMessage}
\alias{print.KafkaMessage}
\title{Print kafka message}
\usage{
\method{print}{KafkaMessage}(x, ...)
}
\arguments{
\item{x}{(`KafkaMessage`) message}

\item{...}{additional args (not used)}
}
\description{
Print kafka message
}
