% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Producer.R
\name{Producer}
\alias{Producer}
\title{Create Kafka Producer}
\usage{
Producer(config)
}
\arguments{
\item{config}{list with configuration options accepted by librdkafka. See https://github.com/confluentinc/librdkafka/blob/master/CONFIGURATION.md for details.}
}
\value{
A kafka producer object. See details for currently available methods
}
\description{
Create Kafka Producer
}
\details{
This function creates an instance of a Kafka Producer based on a list with configuration options. The following methods are currently implemented: ...
}
\examples{
\dontrun{
config <- list(
    "bootstrap.servers" = "my-boostrap-server.com"
)

producer <- Producer(config)

producer$produce(topic = "test-topic", message = "Hello World!")

producer$flush(1000)
}
}
