% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Message.R
\name{result_error_code}
\alias{result_error_code}
\title{Extract error code from result}
\usage{
result_error_code(result)
}
\arguments{
\item{result}{(`KafkaResult`) result from consumer}
}
\description{
Extract error code from result
}
