% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Message.R
\name{result_error_message}
\alias{result_error_message}
\title{Extract error message from result}
\usage{
result_error_message(result)
}
\arguments{
\item{result}{(`KafkaResult`) result from consumer}
}
\description{
Extract error message from result
}
