% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Message.R
\name{result_has_error}
\alias{result_has_error}
\title{Check if result has errors}
\usage{
result_has_error(result)
}
\arguments{
\item{result}{(`KafkaResult`) result from consumer}
}
\description{
Check if result has errors
}
