% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Message.R
\name{result_message}
\alias{result_message}
\title{Extract message from result}
\usage{
result_message(result)
}
\arguments{
\item{result}{(`KafkaResult`) result from consumer}
}
\description{
Extract message from result
}
