% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaMessage.R
\name{print.KafkaMessage}
\alias{print.KafkaMessage}
\title{Print Message}
\usage{
\method{print}{KafkaMessage}(x, ...)
}
\arguments{
\item{x}{a KafkaMessage}

\item{...}{additional arguments. Not used atm}
}
\description{
Print Message
}
